+++ app/components/CustomTextInput.js
