+++ app/components/MissionsView/MissionInfo.js
