+++ app/components/MissionsView/assets/snack-icon.png
