--- app/modules/navigation.js
@@ -1,15 +1,0 @@
-import {NavigationContainer} from '@react-navigation/native';
-import {createStackNavigator} from '@react-navigation/stack';
-
-const Stack = createStackNavigator();
-
-export default function() {
-    return (
-      <NavigationContainer>
-        <Stack.Navigator initialRouteName="login">
-          <Stack.Screen name="login" component={require('../../screens/Login')} />
-          <Stack.Screen name="" component={} />
-        </Stack.Navigator>
-      </NavigationContainer>
-    );
-}
