+++ package.json
@@ -8,18 +8,18 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "expo": "^37.0.7",
-    "react": "~16.9.0",
-    "react-dom": "~16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
+    "react-native-paper": "latest",
+    "react-native-screens": "latest",
+    "react-native-safe-area-context": "latest",
+    "@react-native-community/masked-view": "latest",
     "@react-navigation/native": "^5.1.6",
-    "@react-navigation/stack": "^5.2.11",
+    "@react-navigation/stack": "latest",
     "expo": "~37.0.3",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
     "react-native-elements": "^1.2.7",
-    "react-native-paper": "^3.8.0",
+    "react-native-gesture-handler": "^1.6.1",
     "react-native-web": "~0.11.7"
   },
   "devDependencies": {
